Fixed bug #1230126 and allowed SQL navigation in SQL Sheet

+++ sql12/fw/src/net/sourceforge/squirrel_sql/fw/datasetviewer/cellcomponent/DataTypeLong.java
@@ -115,7 +115,7 @@ public class DataTypeLong
 	 * Neither of the objects is null
 	 */
 	public boolean areEqual(Object obj1, Object obj2) {
-		return ((Long)obj1).equals(obj2);
+		return (obj1).equals(obj2);
 	}
 
 	/*
